requirements.txt
numpy==1.24.1
matplotlib==3.8.0
seaborn==0.13.0
tabulate==0.9.0
psycopg2-binary==2.9.9
pandas==2.1.1